fix: update endpoints still using twitter api
