+++ src/endpoints/quests/twitter_tribe/mod.rs
@@ -1,4 +1,4 @@
-pub mod twitter_name_callback;
 pub mod verify_has_domain;
 pub mod verify_twitter_fw;
+pub mod verify_twitter_name;
 pub mod verify_twitter_rt;
